==> 0_preprocessing.ipynb <==
# Data Preprocessing

## Environment

import pandas as pd
import pyarrow
import re

# mount google drive
from google.colab import drive
drive.mount('/content/gdrive')
%cd /content/gdrive/MyDrive/semantic_song_search

## Load Data

df = pd.read_parquet("data/5mil_song_set.parquet")
df = df.dropna(subset = "lyrics")
df = df.reset_index(drop=True)
print(df.shape)
df.info()

df = df.drop(columns=["features"])
blank = (df["lyrics"] == "")
df = df[~blank]

## Limit by "Good Music"
Started out by limiting by view counts, but then I realized that a lot of people have poor music tastes, especially the masses. So, I printed out a list of artists in my personal collection.

good_artists = ["311", "1975", "Al Green", "Allah-Las", "Animal Collective", "Apples In Stereo", "Arctic Monkeys", "Atlas Sound",
                "Atoms For Peace", "Audioslave", "Autechre", "Average White Band", "Avey Tare", "Art d'Ecco", "A Tribe Called Quest",
                "Art Blakey & The Jazz Messengers", "Barenaked Ladies", "Battles", "Beach Boys", "Beach Fossils", "Bear In Heaven",
                "Beck", "Bert Jansch", "Best Coast", "Billy Idol", "Black Keys", "Black Moth Super Rainbow", "Boards Of Canada",
                "Bob Dylan", "Bodies Of Water", "Bon Iver", "The Books", "Boys Noize", "Brant Bjork", "Brendan Benson",
                "Brightblack Morning Light", "Broken Bells", "The Beatles", "Beastie Boys", "Bob Marley", "Burning Spear",
                "Booker T. & The M.G.'s", "Blur", "Cage The Elephant", "Canned Heat", "Car Seat Headrest", "Caribou", "Carole King",
                "Cat Power", "Cat Stevens", "Chromatics", "Coldplay", "Com Truise", "Counting Crows", "Cream", "Creedence Clearwater Revival",
                "Crystal Castles", "The Crystal Method", "CSS", "Cut Copy", "Childish Gambino", "Charles Mingus", "D.R. Hooker",
                "Dabrye", "Daedelus", "Daft Punk", "Damon Albarn", "Dan Auerbach", "The Dandy Warhols", "Darondo", "David Bowie",
                "The Dead Weather", "Deakin", "Decemberists", "Deep Sea Diver", "Deerhunter", "Department Of Eagles", "Digitalism",
                "Dirty Projectors", "Discovery", "Dispatch", "The Doors", "Dr. Dog", "Danger Mouse", "Das Racist", "Del Tha Funkee Homosapien",
                "Dolly Parton", "The Dandy Warhols", "Dave Brubeck", "Dave Douglas", "Duke Ellington", "Deltron 3030", "DJ Shadow", "DJ Vadim",
                "Doseone", "Dr. Dre", "Dandy Livingstone", "Edward Sharpe & The Magnetic Zeros", "Eels", "Electric President", "Elliott Smith",
                "Elvis Presley", "Endo", "Eric Clapton", "Explosions In The Sky", "Elton John", "Eurythmics", "Eagles", "El-P", "Eminem",
                "Fatboy Slim", "Feist", "Fischerspooner", "Fleet Foxes", "Fleetwood Mac", "Four Tet", "Frank Sinatra", "Frankie Valli",
                "Franz Ferdinand", "Freddie King", "Fu Manchu", "Fugazi", "Future Shuttle", "Frank Sinatra", "Gang Gang Dance", "Garage Voice",
                "Gardens & Villa", "Gary Numan", "Gary Wright", "Gentleman Losers", "Gorillaz", "Green Day", "Greta Van Fleet", "Grizzly Bear",
                "The Guess Who", "George Harrison", "Geto Boys", "Gnarls Barkley", "GZA", "Gregory Isaacs", "Hercules And Love Affair", 
                "Here We Go Magic", "High Places", "Hit Factory", "Hives", "Hot Chip", "Hotel Pools", "Huey Lewis & The News", "Hank Williams",
                "Herbie Hancock", "Handsome Boy Modeling School", "Heems", "Hound Dog Taylor", "Incubus", "Iron & Wine", "Iggy Pop",
                "Idris Muhammad", "Ice Cube", "Jack White", "James Brown", "James Holden", "James Taylor", "Jamie Lidell", "Jane", "Janis Joplin",
                "Jars Of Clay", "Jefferson Airplane", "Jelly Roll Morton", "Jimi Hendrix", "Jon Spencer Blues Explosion", "Junior Boys",
                "Junk Culture", "Justice", "Justin Timberlake", "John Lennon", "Johnny Cash", "J Dilla", "Jurassic 5", "John Lee Hooker",
                "Kelley Polar", "Kills", "Kinks", "Kraftwerk", "Kurt Vile", "Kanye West", "Kid Cudi", "Killer Mike", "Kings of Leon",
                "Late BP Helium", "LCD Soundsystem", "Led Zeppelin", "Leon Bridges", "Les Claypool", "Limp Bizkit", "Lindstrøm", "Linkwood",
                "Loretta Lynn", "Louis Armstrong", "Lil Wayne", "Logic", "Lord T & Eloise", "La's", "Mac DeMarco", "Madonna", "Marcy Playground",
                "Marvin Gaye", "Massive Attack", "Medeski, Martin & Wood", "Memory Cassette", "Memory Tapes", "MGMT", "Michael Jackson", "Moby",
                "Modest Mouse", "MSTRKRFT", "Meters", "Miles Davis", "Mose Allison", "M.I.A", "Mac Miller", "MF DOOM", "Missy Elliott",
                "The Mamas & The Papas", "Neon Indian", "New Order", "Nikka Costa", "Nine Inch Nails", "Nirvana", "Neil Young", "of Montreal",
                "Otis Redding", "Panda Bear", "Pearl Harbor", "Phoenix", "Picastro", "Pinback", "Pink Floyd", "Pixies", "Polvo",
                "The Presidents of the United States of America", "Prince", "Prodigy", "Paul McCartney", "People Under The Stairs", "The Police",
                "Pointer Sisters", "Quarashi", "R.E.M", "Raconteurs", "Radiohead", "Rage Against The Machine", "Ratatat", "Ravi Shankar",
                "Red Hot Chili Peppers", "Royal Blood", "The Rolling Stones", "Ringo Starr", "Ramsey Lewis", "Roots Manuva", "Run The Jewels",
                "Robert Johnson", "Rick James", "Sensational Saints", "Sheryl Crow", "The Shins", "Simian Mobile Disco", "Sonic Youth", "Spoon",
                "St. Vincent", "Steve Miller Band", "Stevie Nicks", "Stevie Wonder", "Sting", "Stone Temple Pilots", "Strokes", "Sublime",
                "Sufjan Stevens", "Surfer Blood", "Sylvan Esso", "Sun Ra", "Shabazz Palaces", "Simon & Garfunkel", "Snail Mail", "Steely Dan", "The Smile", 
                "The Talking Heads", "Tamaryn", "Tame Impala", "Tegan and Sara", "Tenacious D", "Thievery Corporation", "Thom Yorke", "Three Dog Night",
                "Tim Carroll", "Timber Timbre", "Toadies", "TOBACCO", "Tom Petty", "Trentemøller", "The Turtles", "Tycho", "Themselves", "Three 6 Mafia",
                "Thelma Houston", "The Unicorns", "Vince Guaraldi Trio", "Washed Out", "Weeknd", "Weezer", "White Denim", "The White Stripes",
                "The Whitest Boy Alive", "Who", "Widowspeak", "Wolfmother", "Westside Connection", "Yeah Yeah Yeahs", "Yeasayer", "Zero 7", "Zoot Woman"]

print(df.shape)
df = df[df['artist'].isin(good_artists)].reset_index(drop=True)
print(df.shape)

## Profanity

# The Obscenity List - 1600+ popular English profanities and their variations
# https://github.com/surge-ai/profanity
nsfw = pd.read_csv("data/profanity_en.csv")
nsfw[nsfw["text"] == "dumbass"]

df["tag"].value_counts()

from tqdm import tqdm
from tqdm.auto import tqdm
tqdm.pandas()

# Create a regular expression pattern that matches any of the words in nsfw['text']
nsfw_pattern = re.compile('|'.join([re.escape(word) for word in nsfw['text']]))

mask = []
for lyric in tqdm(df['lyrics'], desc='Processing Lyrics'):
    mask.append(nsfw_pattern.search(lyric, re.IGNORECASE) is not None)



# Select only rows without NSFW words
print(df.shape)
mask = pd.Series(mask)
df = df[~mask]
df = df.reset_index(drop=True)
print(df.shape)

df["tag"].value_counts()

## Annotations

# remove annotations (in brackets)
#limit size of brackets to 100 so the regex doesn't stall out on unclosed brackets
df['lyrics'] = df['lyrics'].replace("\[.{,100}]", "", regex = True)

## Reduce Count

desired_count = 17000
fraction = desired_count/df.shape[0]
df = df.sample(frac=fraction, replace=False)
df = df.reset_index(drop=True)

df.shape

## Create Normalized Version of Lyrics (for GloVe)

import nltk
import re
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize

nltk.download('stopwords')
nltk.download('punkt')
nltk.download('wordnet')
nltk.download('averaged_perceptron_tagger')

from nltk.corpus import wordnet
def get_wordnet_pos(treebank_tag):
    if treebank_tag.startswith('J'):
        return wordnet.ADJ
    elif treebank_tag.startswith('V'):
        return wordnet.VERB
    elif treebank_tag.startswith('N'):
        return wordnet.NOUN
    elif treebank_tag.startswith('R'):
        return wordnet.ADV
    else:
        return wordnet.NOUN


def preprocess_text(text):
    # lowercase
    text = text.lower()
    # remove special characters and punctuation
    text = re.sub(r'[^a-zA-Z\s]', ' ', text)
    # tokenize
    tokens = word_tokenize(text)
    # remove stopwords
    stop_words = set(stopwords.words('english'))
    tokens = [word for word in tokens if word not in stop_words]
    # lemmatize
    lemmatizer = WordNetLemmatizer()
    tagged = nltk.pos_tag(tokens)
    tokens = [lemmatizer.lemmatize(word, get_wordnet_pos(tag)) for word, tag in tagged]
    # rejoin string
    preprocessed_text = ' '.join(tokens)
    return preprocessed_text

from tqdm import tqdm
from tqdm.auto import tqdm
tqdm.pandas()

# preprocess text
df['normalized_lyrics'] = df['lyrics'].progress_apply(preprocess_text)

## Review and Save

pd.set_option('max_colwidth', 500)
df.sample(10).T

df.to_parquet("data/5mil_song_set_clean_v2_good_artists.parquet")
