==> 1_lyrics_embed.ipynb <==
# Embeddings

## Environment

!pip install sentence_transformers tiktoken openai -qqq

import numpy as np
import pandas as pd
import pyarrow

import tiktoken
import openai
from openai.embeddings_utils import get_embedding

from sentence_transformers import SentenceTransformer
import torch

from google.colab import drive
drive.mount('/content/gdrive')
%cd /content/gdrive/MyDrive/semantic_song_search

## Load Data

df = pd.read_parquet('data/5mil_song_set_clean_v2_good_artists.parquet')
print(df.shape)
df.info()

pd.set_option('max_colwidth', 500)
df.sample(5).T

df['word_count'] = df['lyrics'].apply(lambda x: len(x.split()))

## Sentence Transformers

# download models

#model_name = 'all-MiniLM-L12-v2'
#model = SentenceTransformer(model_name)
#model.save('./' + model_name)

#model_name = 'all-distilroberta-v1'
#model = SentenceTransformer(model_name)
#model.save('./' + model_name)

# sentence transformer models

minilm = SentenceTransformer('./all-MiniLM-L12-v2')
minilm = minilm.to(torch.device('cuda')) # use GPU

roberta = SentenceTransformer('./all-distilroberta-v1')
roberta = roberta.to(torch.device('cuda')) # use GPU


%%time
embeddings_minilm = minilm.encode(df['lyrics'])

%%time
embeddings_roberta = roberta.encode(df['lyrics'])

# append
df['embedding_minilm'] = list(embeddings_minilm)
df['embedding_roberta'] = list(embeddings_roberta)

## OpenAI Text Embeddings

# openai api config
with open("private/openai_secret_key.txt", "r") as f:
    my_little_secret = f.read()
openai.api_key = my_little_secret

# embedding model parameters
embedding_model = "text-embedding-ada-002"

# check if any songs are too long
encoding = tiktoken.get_encoding("cl100k_base") # encoding for text-embedding-ada-002
max_tokens = 8191  # text-embedding-ada-002 max
sum(df.lyrics.apply(lambda x: len(encoding.encode(x))) > max_tokens)

%%time

from tqdm import tqdm
from tqdm.auto import tqdm
tqdm.pandas()

df["embedding_gpt"] = df.lyrics.progress_apply(lambda x: get_embedding(x, engine=embedding_model))

### GloVe

# download model
#model_name = 'average_word_embeddings_glove.840B.300d'
#model = SentenceTransformer(model_name)
#model.save('./' + model_name)

glove = SentenceTransformer('./average_word_embeddings_glove.840B.300d')

%%time
# compute vector representations of the comments
embeddings_glove = glove.encode(df['normalized_lyrics'])

df['embedding_glove'] = list(embeddings_glove)

### Save for Later

df.to_parquet('data/v2ga_w_embeddings.parquet')
