==> 2_song_search.ipynb <==
# Semantic Search

## Environment

!pip install sentence_transformers tiktoken openai -qqq

import numpy as np
import pandas as pd
import pyarrow

import openai
from openai.embeddings_utils import get_embedding

from sentence_transformers import SentenceTransformer
import torch

from google.colab import drive
drive.mount('/content/gdrive')
%cd /content/gdrive/MyDrive/semantic_song_search

## Load Data

df = pd.read_parquet('data/v2ga_w_embeddings.parquet')
print(df.shape)
df.info()

pd.set_option('max_colwidth', 100)
df.sample(5).T

## Setup Embedding Models

# sentence transformer models
minilm = SentenceTransformer('./all-MiniLM-L12-v2')
roberta = SentenceTransformer('./all-distilroberta-v1')

# openai api config
with open("private/openai_secret_key.txt", "r") as f:
    my_little_secret = f.read()
openai.api_key = my_little_secret
# embedding model
gpt = "text-embedding-ada-002"

# glove
glove = SentenceTransformer('./average_word_embeddings_glove.840B.300d')

## Search

def cosine_similarity(v1, v2):
    dot_product = np.dot(v1, v2)
    v1_norm = np.linalg.norm(v1)
    v2_norm = np.linalg.norm(v2)
    if v1_norm == 0.0 or v2_norm == 0.0:
        return np.nan
    else:
        similarity = dot_product / (v1_norm * v2_norm)
        return similarity

def relevance_scores(query_embed,df,embeddings):
    scores = [cosine_similarity(query_embed, v2) for v2 in df[embeddings]]
    scores = pd.Series(scores)
    return(scores)

def semantic_search(query_text, df = df, model = "glove", return_top = True):
    
    df = df.copy()

    if model == "glove":
      query_embed = glove.encode(query_text)
      scores = relevance_scores(query_embed,df,"embedding_glove")

    if model == "minilm":
      query_embed = minilm.encode(query_text)
      scores = relevance_scores(query_embed,df,"embedding_minilm")

    if model == "roberta":
      query_embed = roberta.encode(query_text)
      scores = relevance_scores(query_embed,df,"embedding_roberta")

    if model == "gpt":
      query_embed = get_embedding(query_text, engine=gpt)
      scores = relevance_scores(query_embed,df,"embedding_gpt")
    
    df['scores'] = scores
    sorted_df = df.sort_values(by = 'scores', ascending = False)

    if return_top:
        result = sorted_df.iloc[0]
        print("\n==============================================\n")
        print(f"Embedding Model: {model}")
        print(f'{result["title"]} by {result["artist"]}\n')
        print(result["lyrics"])
        print("\n==============================================\n")
    else:
        return sorted_df[['title','artist','lyrics','scores']]

pd.set_option('display.width', 100)

## Example

q = "I’m happy sometimes but also sad. You wouldn’t understand"

semantic_search(q, df, "glove", return_top=True)

semantic_search(q, df, "minilm", return_top=True)

semantic_search(q, df, "roberta", return_top=True)

semantic_search(q, df, "gpt", return_top=True)

## Test Loop

def loop_results(q):
  models = ["glove","minilm","roberta","gpt"]
  for model in models:
    semantic_search(q, df, model, return_top=True)

q = "Accept your conditions as they are. Don't struggle with your fate. Find peace in the current state of things."
loop_results(q)
